+++ javascript/UnixLepik.d.ts
@@ -90,6 +90,7 @@ declare class UnixLepik {
      */
     getActiveWindowId(): number;
     setActiveWindow(windowId: string): void;
+    minimizeWindow(windowId: string): void;
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.