+++ javascript/UnixLepik.js
@@ -150,6 +150,10 @@ class UnixLepik {
         const command = `xdotool windowactivate ${windowId}`;
         __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
     }
+    minimizeWindow(windowId) {
+        const command = `xdotool windowminimize ${windowId}`;
+        __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
+    }
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.