+++ javascript/WindowsLepik.d.ts
@@ -91,6 +91,7 @@ declare class WindowsLepik {
     */
     getActiveWindowId(): Promise<number>;
     setActiveWindow(windowId: string): void;
+    minimizeWindow(windowId: string): void;
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.