+++ tests/mouseMove.js
@@ -2,7 +2,7 @@ const lepik = require("../index.js")
 
 async function main() {
   lepik.delay(2000)
-  lepik.setActiveWindow(2426196)
+  lepik.minimizeWindow(2426196)
   lepik.close()
 }
 function sleep(ms) {