+++ typescript/UnixLepik.ts
@@ -168,6 +168,12 @@ class UnixLepik {
     this.#executeShellCommand(command);
   }
 
+  minimizeWindow(windowId: string): void {
+    const command = `xdotool windowminimize ${windowId}`;
+    this.#executeShellCommand(command);
+  }
+
+
   /**
    * Delays the execution for the specified number of milliseconds.
    * @param {number} ms - The number of milliseconds to delay.