+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Full Changelog](https://github.com/borecjeborec1/LepikJS/commits/main)
 
 ##### Implemented enhancements:
+-  minimize window ([#9cb448c](https://github.com/borecjeborec1/LepikJS/commit/9cb448c4e651012ad71f13b9cc423e43b9693e10))
 -  setActiveWindow method ([#7cd3529](https://github.com/borecjeborec1/LepikJS/commit/7cd3529fb8f79dbcc6c8435bbfa38b1e06e3301d))
 -  delay functionality ([#c576ed3](https://github.com/borecjeborec1/LepikJS/commit/c576ed3e53bb839e3e2a85a7eb270178ce36f313)) 
 -  copy and paste methods implementation ([#8240475](https://github.com/borecjeborec1/LepikJS/commit/82404757db63d202bd617728ee1319b49c86fc6b)) 
@@ -13,7 +14,7 @@
 -  v3.0.0 Uses powershell scripts on windows and xdotool on linux machines ([#3d5dc97](https://github.com/Borecjeborec1/LepikJS/commit/3d5dc97f1f5fb773054864fb02e586c8a30793b5)) 
 
 ##### Fixed bugs:
--  wrong content in changelog ([#fd468c0](https://github.com/borecjeborec1/LepikJS/commit/fd468c071040ab1a92eb2586bcd85594b9078cde))
+-  Fixed wrong content in changelog ([#fd468c0](https://github.com/borecjeborec1/LepikJS/commit/fd468c071040ab1a92eb2586bcd85594b9078cde))
 -  Fixed bug where some mouse controlling methods didn't work on certain linux distribuitions. [Issue #10](https://github.com/Borecjeborec1/LepikJS/issues/10). Because of xdotool all methods rely on the X11 . ([#3d5dc97](https://github.com/Borecjeborec1/LepikJS/commit/3d5dc97f1f5fb773054864fb02e586c8a30793b5)) 
 -  Fixed bug where bundled applications didn't work because they required python's bundled exe.  ([#3d5dc97](https://github.com/Borecjeborec1/LepikJS/commit/3d5dc97f1f5fb773054864fb02e586c8a30793b5))